+++ db/schema.rb
@@ -13,7 +13,7 @@
 ActiveRecord::Schema[7.0].define(version: 2023_06_07_132451) do
   create_table "orders", force: :cascade do |t|
     t.string "kind", null: false
-    t.integer "price", null: false
+    t.decimal "price", null: false
     t.string "customer", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
